refactor: move the modal into a separate component
