+++ match-match-game/src/components/header/header.css
@@ -1,5 +1,6 @@
 .header {
   padding-left: 24px;
+  padding-right: 20px;
   width: 100%;
   height: 64px;
   display: flex;
@@ -72,11 +73,13 @@
 .main-button {
   margin-top: 12px;
   margin-left: auto;
-  margin-right: 20px;
   width: 207px;
   height: 36px;
   padding: 10px 16px;
-  background-color: #fff;
+  background-color: #F2F9FE;
+  box-shadow: 0px 3px 5px rgba(0, 0, 0, 0.2),
+              0px 1px 18px rgba(0, 0, 0, 0.12),
+              0px 6px 10px rgba(0, 0, 0, 0.14);
   border-radius: 4px;
   color: #2196F3;
   text-transform: uppercase;
@@ -90,66 +93,4 @@
 .main-button:hover {
   background-color: #1E88E5;
   color: #fff;
-}
-
-.modal {
-  display: none;
-  position: fixed;
-  justify-content: center;
-  align-items: center;
-  z-index: 10;
-  left: 0;
-  top: 0;
-  width: 100%;
-  height: 100%;
-  overflow: auto;
-  background-color: rgba(0,0,0,0.4);
-}
-
-@keyframes modal {
-  from {top: -320px; opacity: 0}
-  to {top: 0; opacity: 1}
-}
-
-.modal__content {
-  position: relative;
-  margin: auto;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  align-items: center;
-  width: 630px;
-  height: 339px;
-  padding: 60px 80px 40px 80px;
-  color: #000;
-  background-color: #fff;
-  box-shadow: 0 4px 8px 0 rgba(0,0,0,0.2), 0 6px 20px 0 rgba(0,0,0,0.19);
-  animation-name: modal;
-  animation-duration: .4s;
-}
-
-.modal__content > form {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  align-items: center;
-  gap: 10px;
-}
-
-input {
-  outline: none;
-}
-
-input:invalid {
-  border: 2px solid red;
-}
-
-input:valid {
-  border: 2px solid black;
-}
-
-.avatar-image {
-  border-radius: 50%;
-  width: 40px;
-  height: 40px;
 }
