+++ match-match-game/src/components/modal/modal.css
@@ -1,0 +1,61 @@
+.modal {
+  display: none;
+  position: fixed;
+  justify-content: center;
+  align-items: center;
+  z-index: 10;
+  left: 0;
+  top: 0;
+  width: 100%;
+  height: 100%;
+  overflow: auto;
+  background-color: rgba(0,0,0,0.4);
+}
+
+@keyframes modal {
+  from {top: -320px; opacity: 0}
+  to {top: 0; opacity: 1}
+}
+
+.modal__content {
+  position: relative;
+  margin: auto;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  align-items: center;
+  width: 630px;
+  height: 339px;
+  padding: 60px 80px 40px 80px;
+  color: #000;
+  background-color: #fff;
+  box-shadow: 0 4px 8px 0 rgba(0,0,0,0.2), 0 6px 20px 0 rgba(0,0,0,0.19);
+  animation-name: modal;
+  animation-duration: .4s;
+}
+
+.modal__content > form {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  align-items: center;
+  gap: 10px;
+}
+
+input {
+  outline: none;
+}
+
+input:invalid {
+  border: 2px solid red;
+}
+
+input:valid {
+  border: 2px solid black;
+}
+
+.avatar-image {
+  border-radius: 50%;
+  width: 40px;
+  height: 40px;
+}
