+++ match-match-game/src/pages/main/main.css
@@ -2,7 +2,7 @@
   display: flex;
   justify-content: center;
   align-items: center;
-  background-color: #E5E5E5;
+  background-color: #fff;
 }
 
 .about-game__content {
